Refactor floor setup into dedicated module
